[libc++][chrono] Adds the sys_info class.
Adds the sys_info class and time_zone::get_info(). The code still
has a few quirks and has not been optimized for performance yet.

The returned sys_info is compared against the output of the zdump tool in
the test giving confidence the implementation is correct.

Implements parts of:
- P0355 Extending <chrono> to Calendars and Time Zones

Implements:
- LWGXXXX The sys_info range should be affected by save